style & test files restructured
